[sw/silicon_creator] Move sigverify.c and unit tests to lib
