+++ sw/device/silicon_creator/lib/sigverify.c
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/silicon_creator/mask_rom/sigverify.h"
+#include "sw/device/silicon_creator/lib/sigverify.h"
 
 #include "sw/device/lib/base/memory.h"
 #include "sw/device/lib/base/mmio.h"
